requirements.txt
python-telegram-bot==20.3
supabase==2.3.4
Flask==2.3.2
